Improve logger error message
For some reason, when using Mirheo as a part of other libraries,
multiple loggers are created. This commit improves the error message
caused by any die() or log-like macros for the cases when the wrong
(uninitialized) logger is called.

Use throw instead of exit(), with exit() we do not get the Python's
stack trace.

Also: silence some gcc warnings.